requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
python-dotenv
requests
openai
celery
redis
python-multipart
alembic
streamlit
httpx
yt-dlp
pillow
opencv-python-headless
